# api_docs.json
{
...
  "info": {
    "description": "API документация для бэкенда Vondic",
    "title": "Vondic Backend API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/posts/": {
      "delete": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "access_token": {
                  "required": true,
                  "type": "string"
                },
                "post_id": {
                  "required": true,
                  "type": "string"
                },
                "user_id": {
                  "required": true,
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Пост удален"
          },
          "400": {
            "description": "Неверные параметры"
          },
          "403": {
            "description": "Нет прав"
          },
          "404": {
            "description": "Пост не найден"
          }
        },
        "summary": "Удалить свой пост (для обычного пользователя)",
        "tags": [
          "Posts"
        ]
      },
      "get": {
        "responses": {
          "200": {
            "description": "Список постов",
            "schema": {
              "items": {
                "properties": {
                  "author_id": {
                    "type": "integer"
                  },
                  "content": {
                    "type": "string"
                  },
                  "id": {
                    "type": "integer"
                  },
                  "title": {
                    "type": "string"
                  }
                },
                "type": "object"
              },
              "type": "array"
            }
          }
        },
        "summary": "Получить список всех постов",
        "tags": [
          "Posts"
        ]
      },
      "post": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "access_token": {
                  "required": true,
                  "type": "string"
                },
                "content": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Пост создан"
          },
          "400": {
            "description": "Ошибка создания"
          }
        },
        "summary": "Создать новый пост",
        "tags": [
          "Posts"
        ]
      },
      "put": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "access_token": {
                  "required": true,
                  "type": "string"
                },
                "content": {
                  "type": "string"
                },
                "post_id": {
                  "required": true,
                  "type": "string"
                },
                "title": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Пост обновлен"
          },
          "400": {
            "description": "Неверные параметры"
          },
          "403": {
            "description": "Нет прав"
          },
          "404": {
            "description": "Пост не найден"
          }
        },
        "summary": "Обновить пост",
        "tags": [
          "Posts"
        ]
      }
    },
...
    "/api/v1/posts/like": {
      "post": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "post_id": {
                  "required": true,
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Лайк добавлен"
          },
          "400": {
            "description": "Неверные параметры"
          },
          "404": {
            "description": "Пост не найден"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Лайкнуть пост",
        "tags": [
          "Posts"
        ]
      }
    },
    "/api/v1/posts/unlike": {
      "post": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "post_id": {
                  "required": true,
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Лайк убран"
          },
          "400": {
            "description": "Неверные параметры"
          },
          "404": {
            "description": "Пост не найден"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Убрать лайк с поста",
        "tags": [
          "Posts"
        ]
      }
    },
    "/api/v1/posts/{post_id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "post_id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Данные поста",
            "schema": {
              "properties": {
                "content": {
                  "type": "string"
                },
                "id": {
                  "type": "integer"
                },
                "title": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "404": {
            "description": "Пост не найден"
          }
        },
        "summary": "Получить пост по ID",
        "tags": [
          "Posts"
        ]
      }
    },
    "/api/v1/posts/{post_id}/comments": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "post_id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Список комментариев",
            "schema": {
              "items": {
                "type": "object"
              },
              "type": "array"
            }
          }
        },
        "summary": "Получить комментарии к посту",
        "tags": [
          "Comments"
        ]
      }
    },
...
  },
  "swagger": "2.0"
}